+++ R/daemon.R
@@ -98,9 +98,9 @@ daemon <- function(
   cv <- cv()
   sock <- socket(if (dispatcher) "poly" else "rep")
   on.exit(reap(sock))
-  autoexit && pipe_notify(sock, cv = cv, remove = TRUE, flag = autoexit)
+  autoexit && pipe_notify(sock, cv, remove = TRUE, flag = autoexit)
   if (length(tls)) tls <- tls_config(client = tls)
-  dial_and_sync_socket(sock, url, asyncdial = asyncdial, tls = tls)
+  dial_and_sync_socket(sock, url, autostart = asyncdial || NA, tls = tls)
 
   `[[<-`(., "sock", sock)
   on.exit(`[[<-`(., "sock", NULL), add = TRUE)
@@ -194,11 +194,11 @@ daemon <- function(
   cv <- cv()
   sock <- socket("rep")
   on.exit(reap(sock))
-  pipe_notify(sock, cv = cv, remove = TRUE)
+  pipe_notify(sock, cv, remove = TRUE)
   dial(sock, url = url, autostart = NA, fail = 2L)
   `[[<-`(., "sock", sock)
   data <- eval_mirai(recv(sock, mode = 1L, block = TRUE))
-  send(sock, data, mode = 1L, block = TRUE) || until(cv, 5000L)
+  send(sock, data, mode = 1L, block = TRUE) || until(cv, .limit_short)
 }
 
 # internals --------------------------------------------------------------------
@@ -225,12 +225,12 @@ eval_mirai <- function(._mirai_.) {
   )
 }
 
-dial_and_sync_socket <- function(sock, url, asyncdial = FALSE, tls = NULL) {
+dial_and_sync_socket <- function(sock, url, autostart = NA, tls = NULL) {
   cv <- cv()
-  pipe_notify(sock, cv = cv, add = TRUE)
-  dial(sock, url = url, autostart = asyncdial || NA, tls = tls, fail = 2L)
+  pipe_notify(sock, cv, add = TRUE)
+  dial(sock, url = url, autostart = autostart, tls = tls, fail = 2L)
   wait(cv)
-  pipe_notify(sock, cv = NULL, add = TRUE)
+  pipe_notify(sock, NULL, add = TRUE)
 }
 
 do_cleanup <- function() {